% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{dce_bootstrap}
\alias{dce_bootstrap}
\title{Generate bootstrap samples based on the probabilistic distribution of attributes of different treatments provided in the dca_samples data}
\usage{
dce_bootstrap(dce_samples, boot_n, na_zero = FALSE, seed = NULL)
}
\arguments{
\item{dce_samples}{samples of probabilistic distribution for the different treatments' attributes. The dataset needs to have the following columns:
Samples (sample number), Attributes (treatments' attributes), Levels (levels of different attributes), and one column per treatment. See dca_sample data in the package as an example of this dataset.}

\item{boot_n}{number of required bootstrap samples}

\item{na_zero}{logical (defaul is FALSE). If true, replace all the NA's in the dataset by zero}

\item{seed}{set seed for reproducibility purposes}
}
\value{
two datasets: 1. generated bootstrap samples, and 2. aggregated probabilistic distribution of attributes of each treatment
}
\description{
Generate bootstrap samples based on the probabilistic distribution of attributes of different treatments provided in the dca_samples data
}
\examples{
dce_boot_obj <- dce_bootstrap(dce_samples = dce_sample, boot_n = 100, seed = 1)
head(dce_boot_obj$boot_data)
head(dce_boot_obj$boot_prob)
}
