% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{dce_CLogitFit}
\alias{dce_CLogitFit}
\title{Generate the bootstrap samples amd fit a conditional logit model}
\usage{
dce_CLogitFit(
  original_data,
  dce_samples,
  dce_sim_data = NULL,
  covariate_list,
  response,
  ID,
  coef_sim_n = 1000,
  trt_list,
  reference_trt = NULL,
  boot_n = NULL,
  na_zero = FALSE,
  seed = NULL
)
}
\arguments{
\item{original_data}{patient data, which need to include the folowing columns: id (patient id), response (a binary outcome), and covariates. See patient_sample data provided in the package for an example of this argument}

\item{dce_samples}{samples of probabilistic distribution for the different treatments' attributes. The dataset needs to have the following columns:
Samples (sample number), Attributes (treatments' attributes), Levels (levels of different attributes), and one column per treatment. See dca_sample data in the package as an example of this dataset.}

\item{dce_sim_data}{generated bootstrape samples of the attributes of different treatments (to estimate the second source of uncertainity in estimated SE's). These data can either be provided by user (e.g., by using the dce_bootstrap function provided in this package), or can be left null (the default value) and then this function will generate a dataset automatically}

\item{covariate_list}{list of covariates' name to be used in the conditional logit model}

\item{response}{response column name}

\item{ID}{id column name}

\item{coef_sim_n}{number of generated samples from the estimated sampling distribution of covariates' effect (to estimate the first source of uncertainity in estimated SE's)}

\item{trt_list}{list of treatment names}

\item{reference_trt}{name of reference treatment in estimating other treatments' uptake. If left null (default), no treatment will be used as the reference treatment.}

\item{boot_n}{number of required bootstrap samples}

\item{na_zero}{logical (defaul is FALSE). If true, replace all the NA's in the dataset by zero}

\item{seed}{set seed for reproducibility purposes}
}
\value{
a summary of the fitted conditional logit model, the estimated utility of each treatment, and the estimated uptake of each treatment relative to the refrence treatment
}
\description{
Generate the bootstrap samples of attributes of different treatments, fit a conditional logit model on the patient data, and estimate effect of each covariates on the outcomes. The function provides SE of the effect estimates that are based on two uncertanities: 1. uncertanity of using sample of patients (not population), and 2. uncertanity in attributes of different treatments.
}
\examples{
dce_CLogit_model <-
  dce_CLogitFit(original_data = patient_sample,
                dce_samples = dce_sample,
                covariate_list = c("X1_1", "X1_2", "X2_1", "X2_2", "X2_3"),
                response = "response",
                ID = "id",
                coef_sim_n = 100,
                trt_list = list(trt1 = "X1_1 + X1_2 + X2_1",
                                trt2 = "X1_1 + X2_1 + X2_2 + X2_3"),
                boot_n = 100, seed = 1)
dce_CLogit_model$model_summary
dce_CLogit_model$coefficient_estimates
dce_CLogit_model$reference_TRT_utility
dce_CLogit_model$`trt1 utility`
dce_CLogit_model$`trt1 uptake`

}
