% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\docType{data}
\name{dce_sample}
\alias{dce_sample}
\title{DCE samples}
\format{
\subsection{\code{dce_sample}}{

A data frame with 14 rows and 8 columns:
\describe{
\item{Samples}{sample number}
\item{Attributes}{attributes of each treatment}
\item{Levels}{levels of each attribute}
\item{no trt, trt1, trt2, trt3, trt4}{treatments}
}
}
}
\source{
simulated data
}
\usage{
dce_sample
}
\description{
A simulated sample of DCE data
}
\keyword{datasets}
