% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\docType{data}
\name{patient_sample}
\alias{patient_sample}
\title{DCE samples}
\format{
\subsection{\code{patient_sample}}{

A data frame with 210 rows and 9 columns:
\describe{
\item{id}{patient id}
\item{response}{data outcomes}
\item{X1, X2}{covariates with their original values}
\item{X1_1. X1_2, X2_1, X2_2, X2_3}{covariates converted to their corresponding dummy variables}
}
}
}
\source{
simulated data
}
\usage{
patient_sample
}
\description{
A simulated sample of patient data for a DCE analysis
}
\keyword{datasets}
